requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure_storage==0.37.0
isodate==0.6.1
psycopg2==2.9.5
psycopg2_binary==2.9.4